[5.17] Добавил фоны, вынес логику в хук
