+++ packages/client/src/pages/Main/SnakeGame/SnakeGame.module.scss
@@ -1,0 +1,19 @@
+.snakeGame {
+    position: relative;
+
+    &__score {
+        position: absolute;
+        right: 10px;
+        top: 10px;
+        padding: 5px;
+        background: rgb(0,0,0,0.5);
+        color: orange;
+        z-index: 1;
+    }
+
+    &__canvas {
+        display: block;
+        position: absolute;
+        inset: 0;
+    }
+}
